+++ src/stub/java/io/lumine/xikage/mythicmobs/adapters/AbstractItemStack.java
@@ -1,0 +1,4 @@
+package io.lumine.xikage.mythicmobs.adapters;
+
+public interface AbstractItemStack {
+}