+++ src/stub/java/io/lumine/xikage/mythicmobs/adapters/bukkit/BukkitItemStack.java
@@ -1,0 +1,9 @@
+package io.lumine.xikage.mythicmobs.adapters.bukkit;
+
+import io.lumine.xikage.mythicmobs.adapters.AbstractItemStack;
+import org.bukkit.inventory.ItemStack;
+
+public class BukkitItemStack implements AbstractItemStack {
+    public BukkitItemStack(ItemStack item) {
+    }
+}