+++ src/stub/java/io/lumine/xikage/mythicmobs/drops/DropMetadata.java
@@ -1,0 +1,4 @@
+package io.lumine.xikage.mythicmobs.drops;
+
+public class DropMetadata {
+}