+++ src/stub/java/io/lumine/xikage/mythicmobs/drops/IItemDrop.java
@@ -1,0 +1,7 @@
+package io.lumine.xikage.mythicmobs.drops;
+
+import io.lumine.xikage.mythicmobs.adapters.AbstractItemStack;
+
+public interface IItemDrop {
+    AbstractItemStack getDrop(DropMetadata metadata);
+}