+++ src/stub/java/io/lumine/xikage/mythicmobs/drops/droppables/CustomDrop.java
@@ -1,0 +1,7 @@
+package io.lumine.xikage.mythicmobs.drops.droppables;
+
+public class CustomDrop {
+    public String getConfigLine() {
+        return "";
+    }
+}