+++ src/stub/java/io/lumine/xikage/mythicmobs/io/MythicLineConfig.java
@@ -1,0 +1,4 @@
+package io.lumine.xikage.mythicmobs.io;
+
+public class MythicLineConfig {
+}